# services/app/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nisaba app service",
    "version": "0.1.0"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MarkInput": {
        "type": "object",
        "required": [
          "start",
          "end",
          "kind",
          "author",
          "timestamp"
        ],
        "properties": {
          "id": {
            "type": [
              "integer",
              "null"
            ]
          },
          "start": {
            "type": "integer",
            "format": "int32"
          },
          "end": {
            "type": "integer",
            "format": "int32"
          },
          "kind": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CompileRequest": {
        "type": "object",
        "required": [
          "project_id",
          "entry",
          "sources",
          "view"
        ],
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid"
          },
          "entry": {
            "type": "string",
            "description": "Entrypoint path; must exist in sources"
          },
          "sources": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Path \u2192 Typst source map"
          },
          "marks": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/MarkInput"
              }
            }
          },
          "view": {
            "type": "string",
            "enum": [
              "baseline",
              "proposed",
              "redline",
              "public"
            ]
          }
        }
      },
      "CompileResponse": {
        "type": "object",
        "required": [
          "build_id"
        ],
        "properties": {
          "pdf_base64": {
            "type": [
              "string",
              "null"
            ],
            "format": "byte"
          },
          "span_map": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "diagnostics": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "outline": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "build_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/compile": {
      "post": {
        "operationId": "compile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Compiled PDF + diagnostics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResponse"
                }
              }
            }
          }
        }
      }
    }
  }
}